+++ src/legacy/cxgo_client__gui__servopts__guiserv.go
@@ -572,7 +572,7 @@ func nox_xxx_guiServerOptionsProcPre_4585D0(win *gui.Window, a2, a3, a4 uintptr)
 				sub_46ACE0(dword_5d4594_1046492, 10141, 10141, 1)
 			} else {
 				nox_window_set_hidden((*gui.Window)(unsafe.Pointer(v12)), 1)
-				dword_5d4594_1046532 = uint32(nox_xxx_guiServerPlayersLoad_456270(dword_5d4594_1046492))
+				dword_5d4594_1046532 = nox_xxx_guiServerPlayersLoad_456270(dword_5d4594_1046492)
 				sub_46ACE0(dword_5d4594_1046492, 10141, 10141, 1)
 			}
 			nox_window_set_hidden((*gui.Window)(dword_5d4594_1046524), 1)
